ex1/requirements.txt
anyio 
appnope 
argon2-cffi 
asttokens 
attrs 
backcall 
beautifulsoup4 
bleach 
cffi 
comm 
contourpy==1.2.0
cycler==0.12.1
decorator 
defusedxml 
entrypoints 
exceptiongroup 
executing 
fastjsonschema 
fonttools==4.47.0
idna 
importlib-resources 
ipykernel 
ipython 
ipython-genutils==0.2.0
ipywidgets 
jedi 
Jinja2 
jsonschema 
jsonschema-specifications 
jupyter 
jupyter-console 
jupyter-server 
jupyter_client 
jupyter_core 
jupyterlab-widgets 
jupyterlab_pygments 
kiwisolver==1.4.5
MarkupSafe 
matplotlib==3.8.2
matplotlib-inline 
mediapy==1.2.0
mistune 
nbclassic 
nbclient 
nbconvert 
nbformat 
nest-asyncio 
notebook 
notebook_shim 
numpy 
packaging 
pandocfilters 
parso 
pexpect 
pickleshare 
Pillow 
pkgutil_resolve_name 
platformdirs 
prometheus-client 
prompt-toolkit 
ptyprocess 
pure-eval 
pycparser 
Pygments 
pyparsing==3.1.1
python-dateutil 
pyzmq 
qtconsole 
QtPy 
referencing 
rpds-py 
Send2Trash 
six 
sniffio 
soupsieve 
stack-data 
terminado 
tinycss2 
tornado 
traitlets 
typing_extensions 
wcwidth 
webencodings 
websocket-client 
widgetsnbextension 
zipp 
